+++ app/src/main/java/com/koreatech/naeilro/ui/facility/FacilityFragment.java
@@ -1,4 +1,4 @@
-package com.koreatech.naeilro.house;
+package com.koreatech.naeilro.ui.facility;
 
 import android.os.Bundle;
 
@@ -11,12 +11,11 @@ import android.view.ViewGroup;
 import com.koreatech.naeilro.R;
 
 
-public class HouseFragment extends Fragment {
-
+public class FacilityFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_house, container, false);
+        return inflater.inflate(R.layout.fragment_facility, container, false);
     }
 }
