+++ app/src/main/java/com/koreatech/naeilro/ui/house/HouseFragment.java
@@ -1,4 +1,4 @@
-package com.koreatech.naeilro.weather;
+package com.koreatech.naeilro.ui.house;
 
 import android.os.Bundle;
 
@@ -11,11 +11,12 @@ import android.view.ViewGroup;
 import com.koreatech.naeilro.R;
 
 
-public class WeatherFragment extends Fragment {
+public class HouseFragment extends Fragment {
+
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_weather, container, false);
+        return inflater.inflate(R.layout.fragment_house, container, false);
     }
 }
