+++ app/src/main/java/com/koreatech/naeilro/ui/reports/ReportsFragment.java
@@ -1,4 +1,4 @@
-package com.koreatech.naeilro.train;
+package com.koreatech.naeilro.ui.reports;
 
 import android.os.Bundle;
 
@@ -11,12 +11,11 @@ import android.view.ViewGroup;
 import com.koreatech.naeilro.R;
 
 
-public class TrainInfoFragment extends Fragment {
-
+public class ReportsFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_train_info, container, false);
+        return inflater.inflate(R.layout.fragment_reports, container, false);
     }
 }
