+++ app/src/main/res/layout/fragment_facility.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".facility.FacilityFragment">
+    tools:context=".ui.facility.FacilityFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
