+++ app/src/main/res/layout/fragment_festival.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".festival.FestivalFragment">
+    tools:context=".ui.festival.FestivalFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
