+++ app/src/main/res/layout/fragment_home.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".HomeFragment" >
+    tools:context=".ui.main.HomeFragment" >
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
