+++ app/src/main/res/layout/fragment_my_info.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyInfoFragment">
+    tools:context=".ui.main.MyInfoFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
