+++ app/src/main/res/layout/fragment_my_plan.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyPlanFragment" >
+    tools:context=".ui.main.MyPlanFragment" >
 
     <TextView
         android:id="@+id/text_notifications"
