+++ app/src/main/res/layout/fragment_shared_info.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SharedInfoFragment" >
+    tools:context=".ui.main.SharedInfoFragment" >
 
     <TextView
         android:id="@+id/text_dashboard"
