+++ app/src/main/res/layout/fragment_tour_spot.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".tourspot.TourSpotFragment">
+    tools:context=".ui.tourspot.TourSpotFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
