+++ app/src/main/res/layout/fragment_train_info.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".train.TrainInfoFragment">
+    tools:context=".ui.train.TrainInfoFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
