+++ app/src/main/res/layout/fragment_weather.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".weather.WeatherFragment">
+    tools:context=".ui.weather.WeatherFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
